requirements.txt
Django>=4.2.11
djangorestframework>=3.15.1
celery>=5.3.4
django-redis>=5.4.0
django-shortuuidfield>=0.1.3
psutil>=5.9.5
PyMySQL>=1.1.0
kombu>=5.3.2
poc-tool>=1.2.1
djangorestframework-simplejwt>=5.3.1
drf-yasg>=1.21.7
pocsuite3>=2.0.8
eventlet>=0.33
flower>=2.0.1
django-db-connection-pool>=1.2.5
mmh3>=4.0.1
pyOpenSSL>=23.2.0
openpyxl>=3.1.2
beautifulsoup4>=4.12.3
aiohttp>=3.9.3
aiodns>=3.2.0
pandas>=2.2.2